[fix-new-pools] refactoring blockchainAPI moved to BlockchainService, updated packages
