+++ package.json
@@ -67,16 +67,16 @@
     "@storybook/react": "^5.3.18",
     "@svgr/webpack": "4.3.3",
     "@types/enzyme": "^3.10.8",
-    "@types/jest": "^26.0.22",
+    "@types/jest": "25.1.3",
     "@types/react": "^17.0.3",
     "@types/react-dom": "^17.0.3",
     "@types/react-router": "^5.1.13",
     "@types/react-router-dom": "^5.1.7",
     "@typescript-eslint/eslint-plugin": "^2.31.0",
     "@typescript-eslint/parser": "^2.31.0",
-    "@emotion/jest": "11.0.0",
+    "@emotion/jest": "11.3.0",
     "babel-eslint": "^10.1.0",
-    "babel-jest": "^24.9.0",
+    "babel-jest": "27.0.2",
     "babel-loader": "8.1.0",
     "babel-plugin-named-asset-import": "^0.3.6",
     "css-loader": "3.4.2",
@@ -98,7 +98,7 @@
     "html-webpack-plugin": "4.0.0-beta.11",
     "husky": "^4.2.5",
     "image-webpack-loader": "^7.0.1",
-    "jest": "^24.9.0",
+    "jest": "27.0.3",
     "jest-environment-jsdom-fourteen": "1.0.1",
     "jest-resolve": "24.9.0",
     "jest-watch-typeahead": "0.4.2",
@@ -119,7 +119,7 @@
     "styled-components": "^5.2.0",
     "styled-reset": "^4.2.3",
     "ts-node": "^8.10.1",
-    "typescript": "^3.8.3",
+    "typescript": "4.2.4",
     "webpack": "^4.43.0",
     "webpack-cli": "^3.3.11",
     "webpack-dev-server": "^3.10.3"
